requirements.txt
Flask
scikit-learn
joblib
gunicorn
Pillow==10.0.0
tensorflow
flask-cors